temp work to add python API to make it easy to deform objects with compute shaders
The goal is to add the possibility for an user to create a SSBO of
positions aligned on vertex numbers for an evaluted object.

He can then set the positions as he wants in the compute shader.

Then he can call an helper function to update render cache vbo::Position
and vbo::CornerNormal with a precompiled compute shader.

This technique was used for gpu skinning in upbge:

https://github.com/UPBGE/upbge/blob/master/source/gameengine/Converter/BL_ArmatureObject.cpp